IMU offsets

+++ robot/src/workspace/src/kapibara_vel_saltis_bridge/include/kapibara_vel_saltis_bridge/datatypes.h
@@ -28,7 +28,7 @@ typedef struct aux_cfg
 
 typedef struct motor_cfg
 {
-    int16_t velocities[2];
+    int32_t velocities[2];
 } motor_cfg_t ;
 
 typedef struct servo_cfg
